# openapi/assetBookingManagementOpenAPISpec.yaml
openapi: 3.1.0
info:
  title: Asset Booking Management
  description: "This is the OpenAPI documentation for documenting and testing endpoints of the Asset Booking Management Project.
    \ The Asset Booking Manager API allows users to discover, reserve, and manage
    \ shared assets such as meeting rooms, vehicles, devices, and other equipment
    \ within an organization. It exposes endpoints for searching assets by type
    \ and availability, creating and updating reservations, and enforcing booking
    \ rules and time windows for different asset categories. Consumers can
    \ integrate this API into internal portals or mobile apps to streamline
    \ scheduling, prevent double-bookings, and improve overall asset utilization
    \ through centralized, auditable booking workflows."
  contact:
    name: Mladen Banović
    url: https://www.maurer-electronics.hr/
    email: mladen.banovic@maurer-electronics.hr
  version: "1.0"
servers:
- url: http://127.0.0.1:8080/
  description: Development
...
paths:
...
  /v1/bookings:
    get:
      tags:
      - Bookings
      summary: Read list of bookings
      description: Only available to authenticated users.
      operationId: getAll_1
      parameters:
      - name: filter
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/BookingFilter"
      - name: page
        in: query
        description: One-based page index (1..N)
        required: false
        schema:
          type: integer
          default: 1
          minimum: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          type: integer
          default: 50
          minimum: 1
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          items:
            type: string
      responses:
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "409":
          description: Conflict
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "422":
          description: Unprocessable Content
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageBookingResponseDTO"
      security:
      - Bearer Authentication: []
    post:
      tags:
      - Bookings
      summary: Create a booking
      description: Only available to authenticated users.
      operationId: create_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingCreateDTO"
        required: true
      responses:
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "409":
          description: Conflict
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "422":
          description: Unprocessable Content
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BookingResponseDTO"
      security:
      - Bearer Authentication: []
  /v1/booking/recurring:
    post:
      tags:
      - Bookings
      summary: Create recurring bookings
      description: Creates multiple independent bookings at once. All slots must be in the same month. Only available to authenticated users.
      operationId: createRecurring
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecurringBookingCreateDTO"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookingResponseDTO"
        "404":
          description: Not Found (User or Asset not found)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "400":
          description: Bad Request (Validation failed, e.g., dates not in the same month)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "409":
          description: Conflict (One or more time slots overlap with existing bookings)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "422":
          description: Unprocessable Content
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security:
      - Bearer Authentication: []
  /v1/bookings/{id}/reject:
    post:
      tags:
      - Bookings
      summary: Reject a pending booking
      description: "Only available to the manager of the user who created the booking,\
        \ or an ADMIN."
      operationId: reject
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "409":
          description: Conflict
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "422":
          description: Unprocessable Content
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BookingResponseDTO"
      security:
      - Bearer Authentication: []
  /v1/bookings/{id}/approve:
    post:
      tags:
      - Bookings
      summary: Approve a pending booking
      description: "Only available to the manager of the user who created the booking,\
        \ or an ADMIN."
      operationId: approve
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "409":
          description: Conflict
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "422":
          description: Unprocessable Content
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BookingResponseDTO"
      security:
      - Bearer Authentication: []
...
  /v1/bookings/{id}:
    get:
      tags:
      - Bookings
      summary: Read booking by ID
      description: Only available to authenticated users.
      operationId: getById_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "409":
          description: Conflict
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "422":
          description: Unprocessable Content
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BookingResponseDTO"
      security:
      - Bearer Authentication: []
    patch:
      tags:
      - Bookings
      summary: Update booking
      description: "Only available to users with role: ADMIN."
      operationId: update_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingUpdateDTO"
        required: true
      responses:
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "409":
          description: Conflict
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "422":
          description: Unprocessable Content
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BookingResponseDTO"
      security:
      - Bearer Authentication: []
...
components:
  schemas:
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        properties:
          type: object
          additionalProperties: {}
...
    BookingCreateDTO:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        assetId:
          type: integer
          format: int64
        bookingStart:
          type: string
          format: date-time
        bookingEnd:
          type: string
          format: date-time
        notes:
          type: string
          maxLength: 1000
          minLength: 0
      required:
      - assetId
      - bookingEnd
      - bookingStart
      - userId
    TimeSlotDTO:
      type: object
      properties:
        bookingStart:
          type: string
          format: date-time
        bookingEnd:
          type: string
          format: date-time
      required:
      - bookingEnd
      - bookingStart
    RecurringBookingCreateDTO:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        assetId:
          type: integer
          format: int64
        timeSlots:
          type: array
          minItems: 1
          maxItems: 31
          items:
            $ref: "#/components/schemas/TimeSlotDTO"
        notes:
          type: string
          maxLength: 1000
          minLength: 0
      required:
      - assetId
      - timeSlots
      - userId
    AssetSummaryDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        category:
          $ref: "#/components/schemas/CategorySummaryDTO"
        status:
          type: string
          enum:
          - ACTIVE
          - INACTIVE
          - DAMAGED
          - DELETED
        description:
          type: string
        location:
          type: string
    BookingResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        user:
          $ref: "#/components/schemas/UserSummaryDTO"
        asset:
          $ref: "#/components/schemas/AssetSummaryDTO"
        status:
          type: string
          enum:
          - PENDING
          - APPROVED
          - CANCELLED
          - REJECTED
          - COMPLETED
        bookingStart:
          type: string
          format: date-time
        bookingEnd:
          type: string
          format: date-time
        notes:
          type: string
    CategorySummaryDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        bookingPeriod:
          type: string
          enum:
          - HOUR
          - DAY
          - WEEK
          - MONTH
        approval:
          type: boolean
    UserSummaryDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        surname:
          type: string
        email:
          type: string
        role:
          type: string
          enum:
          - EMPLOYEE
          - MANAGER
          - ADMIN
        managerEmail:
          type: string
...
    BookingUpdateDTO:
      type: object
      properties:
        status:
          type: string
          enum:
          - PENDING
          - APPROVED
          - CANCELLED
          - REJECTED
          - COMPLETED
        bookingStart:
          type: string
          format: date-time
        bookingEnd:
          type: string
          format: date-time
        notes:
          type: string
          maxLength: 1000
          minLength: 0
...
    PageableObject:
      type: object
      properties:
        unpaged:
          type: boolean
        paged:
          type: boolean
        pageSize:
          type: integer
          format: int32
        pageNumber:
          type: integer
          format: int32
        offset:
          type: integer
          format: int64
        sort:
          $ref: "#/components/schemas/SortObject"
    SortObject:
      type: object
      properties:
        unsorted:
          type: boolean
        sorted:
          type: boolean
        empty:
          type: boolean
...
    BookingFilter:
      type: object
      properties:
        status:
          type: string
          enum:
          - PENDING
          - APPROVED
          - CANCELLED
          - REJECTED
          - COMPLETED
        userId:
          type: integer
          format: int64
        assetId:
          type: integer
          format: int64
        categoryId:
          type: integer
          format: int64
        bookingStart:
          type: string
          format: date-time
        bookingEnd:
          type: string
          format: date-time
    PageBookingResponseDTO:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject"
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/BookingResponseDTO"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        empty:
          type: boolean
...
  securitySchemes:
    Bearer Authentication:
      type: http
      scheme: bearer
      bearerFormat: JWT